requirements.txt.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
PyJWT==2.8.0
bcrypt==4.1.3
gunicorn==22.0.0
